+++ src/Renderer/renderBuffer.cc
@@ -41,7 +41,7 @@ void fillRenderBufferWithGivenMat(Renderer &r, Material &m){
     u32 submittedQuads = 0;
     for(u32 x=0; x<m.registeredEntities.count; x+=1){
 	Entity e = m.registeredEntities[x];
-	auto *transform = (Component::Transform*)getComponent(e, (u32)ComponentID::TRANSFORM);
+	auto *transform = (Component::Transform*)getComponent(e, ComponentID::TRANSFORM);
 	if(transform == nullptr){continue;};
 	submitQuad(r, transform->genMatrix(), m);
 	submittedQuads += 1;
